requirements.txt
python-decouple==3.8
psycopg2-binary==2.9.5
bcrypt==4.0.1
pydantic==1.9.1
typing-extensions==4.3.0
django-ninja==0.22.2
django-ninja-extra==0.19.1
uvicorn==0.20.0
requests==2.28.1